% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batter_pitcher_ids}
\alias{batter_pitcher_ids}
\title{MLB Batter-Pitcher IDs}
\format{
A data frame with variables:
\describe{
\item{playerid}{Player ID matching Fangraphs}
\item{Season}{MLB season year}
}
}
\source{
Fangraphs via baseballr package
}
\usage{
batter_pitcher_ids
}
\description{
A dataset identifying players who were both batters and pitchers in a season.
}
\keyword{datasets}
