% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhl_birthdays}
\alias{nhl_birthdays}
\title{NHL Player Birthdays}
\format{
A data frame with variables:
\describe{
\item{...1}{Row index}
\item{id}{Player ID matching the NHL API}
\item{birthDate}{Date of birth in YYYY-MM-DD format}
}
}
\source{
NHL API
}
\usage{
nhl_birthdays
}
\description{
A dataset containing birth dates for NHL players.
}
\keyword{datasets}
