% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwhl_players_with_ages}
\alias{pwhl_players_with_ages}
\title{PWHL Player Ages}
\format{
A data frame with variables:
\describe{
\item{player_id}{Player ID matching the fastRhockey package}
\item{birth_year}{Year of birth}
\item{player_name}{Player's full name}
}
}
\source{
EliteProspects web scraping
}
\usage{
pwhl_players_with_ages
}
\description{
A dataset containing birth years for PWHL players.
}
\keyword{datasets}
